Tag does not get saved
Important to save the tags as user would want their tags to remain after closing the app

Update TaskFile to save the tags and fix some additional bugs. Add function to search tasks using tags.

This allows for users to retain the tags saved and search for tasks based on tags.